Update policy for Subscription model

+++ app/Http/Controllers/SubscriptionController.php
@@ -14,7 +14,6 @@ class SubscriptionController extends Controller
     // index subscriptions
     public function index()
     {
-        dd("subs index policy");
         $this->authorize('viewAny', Subscription::class);
         $subscriptions = Subscription::all();
         return view('subscriptions.index', compact('subscriptions'));
@@ -24,12 +23,14 @@ class SubscriptionController extends Controller
     //show subscriptions create form
     public function create()
     {
+        $this->authorize('create', Subscription::class);
         return view('subscriptions.create');
     }
 
     // store subscriptions
     public function store()
     {
+        $this->authorize('create', Subscription::class);
         Subscription::create(request()->validate([
             'plan' => 'required',
             'cost_percentage' => 'required'
@@ -41,18 +42,21 @@ class SubscriptionController extends Controller
     // show a single subscription
     public function show(Subscription $subscription)
     {
+        $this->authorize('view', $subscription);
         return view('subscriptions.show', compact('subscription'));
     }
 
     //edit form is available
     public function edit(Subscription $subscription)
     {
+        $this->authorize('update', $subscription);
         return view('subscriptions.edit', compact('subscription'));
     }
 
     // update a subscription
     public function update(Subscription $subscription)
     {
+        $this->authorize('update', $subscription);
         $data = request()->validate([
             'plan' => 'required',
             'cost_percentage' => 'required|integer'
@@ -64,6 +68,7 @@ class SubscriptionController extends Controller
     //delete a subscription
     public function destroy(Subscription $subscription)
     {
+        $this->authorize('delete', $subscription);
         $subscription->delete();
         return redirect()->route('subscriptions.index');
     }
